python: Break circular dependency
Move the features checking logic to a dedicated file.
