+++ src/python/espressomd/collision_detection.py
@@ -1,3 +1,4 @@
+#
 # Copyright (C) 2010-2022 The ESPResSo project
 #
 # This file is part of ESPResSo.
@@ -14,9 +15,11 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+
 from .script_interface import ScriptInterfaceHelper, script_interface_register
 from .interactions import BondedInteraction, BondedInteractions
-from .__init__ import assert_features
+from .code_features import assert_features
 
 
 @script_interface_register
