+++ src/python/espressomd/ekboundaries.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from .script_interface import script_interface_register
-from .__init__ import has_features
+from .code_features import has_features
 import espressomd.lbboundaries
