+++ src/python/espressomd/virtual_sites.py
@@ -1,3 +1,4 @@
+#
 # Copyright (C) 2010-2022 The ESPResSo project
 #
 # This file is part of ESPResSo.
@@ -14,7 +15,9 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from .__init__ import has_features
+#
+
+from .code_features import has_features
 from .script_interface import ScriptInterfaceHelper, script_interface_register
 
 if has_features("VIRTUAL_SITES"):
